requirements.txt
# Dependencias principales
tensorflow>=2.6.0
tensorflow-transform>=1.4.0
tensorflow-model-analysis>=0.42.0
tfx>=1.10.0
confluent-kafka>=1.9.2
sgp4>=2.22
googleapis-common-protos>=1.56.4
ortools>=9.5.2237

# API y servicios web
fastapi>=0.95.1
uvicorn[standard]>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0

# Visualización y frontend
streamlit>=1.22.0
pydeck>=0.8.0
matplotlib>=3.7.1

# Análisis de datos y ML
networkx>=3.1
pandas>=2.0.1
numpy>=1.24.3
stellargraph>=1.2.1

# Testing y desarrollo
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
